requirements.txt
openai
streamlit-audiorec
openai-whisper
python-dotenv
pillow